requirements.txt
Pillow
requests
beautifulsoup4
ddddocr
python-dotenv
colorlog
flask
flask-cors
